webserver/requirements.txt
flask
jsonify
requests
gunicorn
pika
werkzeug
flask_sqlalchemy
Flask-Migrate
pylint-flask
pylint-flask-sqlalchemy
psycopg2
pillow
imagehash
redis
prometheus-flask-exporter
